requirements.txt
future
numpy
opencv-python
Pillow
ptflops
thop
tqdm
tensorboard
tensorboardX
bytecode
einops